# backend/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "LeadLearnHub API",
    "description": "API documentation for LeadLearnHub backend services",
    "version": "1.0.0",
    "contact": {
      "name": "LeadLearnHub Support",
      "url": "https://leadlearnhub.vercel.app"
    },
    "license": {
      "name": "MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:4000",
      "description": "Local development server"
    },
    {
      "url": "https://leadlearnhub.vercel.app",
      "description": "Production server"
    }
  ],
...
  "paths": {
...
    "/api/profiles/me": {
      "get": {
        "tags": ["Profiles"],
        "summary": "Get current user profile (with specific role profile)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "User profile",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          }
        }
      }
    },
    "/api/profiles/student": {
      "patch": {
        "tags": ["Profiles"],
        "summary": "Update student profile",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StudentProfile"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated user",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          }
        }
      }
    },
    "/api/profiles/tutor": {
      "patch": {
        "tags": ["Profiles"],
        "summary": "Update tutor profile",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TutorProfile"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated user",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          }
        }
      }
    },
    "/api/profiles/school": {
      "patch": {
        "tags": ["Profiles"],
        "summary": "Update school profile",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SchoolProfile"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated user",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          }
        }
      }
    },
    "/api/profiles/upload-avatar": {
      "post": {
        "tags": ["Profiles"],
        "summary": "Upload avatar/logo placeholder",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Avatar URL",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "avatarUrl": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "clxyz123"
          },
          "email": {
            "type": "string",
            "example": "user@example.com"
          },
          "firstName": {
            "type": "string",
            "example": "John"
          },
          "lastName": {
            "type": "string",
            "example": "Doe"
          },
          "phone": {
            "type": "string",
            "nullable": true,
            "example": null
          },
          "avatarUrl": {
            "type": "string",
            "nullable": true,
            "example": null
          },
          "role": {
            "type": "string",
            "enum": ["STUDENT", "TUTOR", "SCHOOL", "ADMIN"],
            "example": "STUDENT"
          },
          "isVerified": {
            "type": "boolean",
            "example": false
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "student": {
            "$ref": "#/components/schemas/StudentProfile"
          },
          "tutor": {
            "$ref": "#/components/schemas/TutorProfile"
          },
          "school": {
            "$ref": "#/components/schemas/SchoolProfile"
          }
        }
      },
      "StudentProfile": {
        "type": "object",
        "properties": {
          "classLevel": {
            "type": "string"
          },
          "schoolName": {
            "type": "string"
          },
          "age": {
            "type": "integer"
          },
          "gender": {
            "type": "string",
            "enum": ["MALE", "FEMALE", "OTHER"]
          },
          "avatar": {
            "type": "string"
          }
        }
      },
      "TutorProfile": {
        "type": "object",
        "properties": {
          "subjects": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "qualifications": {
            "type": "string"
          },
          "bio": {
            "type": "string"
          },
          "experience": {
            "type": "integer"
          },
          "avatar": {
            "type": "string"
          }
        }
      },
      "SchoolProfile": {
        "type": "object",
        "properties": {
          "schoolName": {
            "type": "string"
          },
          "address": {
            "type": "string"
          },
          "schoolType": {
            "type": "string"
          },
          "logo": {
            "type": "string"
          }
        }
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT authentication token"
      }
    }
  }
}